+++ testing/racker/test_run.py
@@ -29,19 +29,19 @@ def test_run_image_invalid(caplog):
 
 def test_run_command_docker_image_success(capfd, delay):
     """
-    Spawn a Debian 10 "buster" container and run `hostnamectl` on it.
+    Spawn a Debian container and run `hostnamectl` on it.
     The container image will be acquired from Docker.
     """
     runner = CliRunner()
 
-    result = runner.invoke(cli, "run -it --rm debian:buster-slim /usr/bin/hostnamectl", catch_exceptions=False)
-    assert result.exit_code == 0
+    result = runner.invoke(cli, "run -it --rm debian:trixie-slim /usr/bin/hostnamectl", catch_exceptions=False)
+    assert result.exit_code == 0, result.stderr
 
     captured = capfd.readouterr()
 
     assert "Static hostname: debuerreotype" in captured.out
     assert "Virtualization: systemd-nspawn" in captured.out
-    assert "Operating System: Debian GNU/Linux 10 (buster)" in captured.out
+    assert "Operating System: Debian GNU/Linux 13 (trixie)" in captured.out
     assert captured.err == ""
 
 
@@ -53,7 +53,7 @@ def test_run_command_curated_cloudimage_success(capfd, delay):
     runner = CliRunner()
 
     result = runner.invoke(cli, "run -it --rm ubuntu-jammy /usr/bin/hostnamectl", catch_exceptions=False)
-    assert result.exit_code == 0
+    assert result.exit_code == 0, result.stderr
 
     captured = capfd.readouterr()
 
@@ -63,19 +63,15 @@ def test_run_command_curated_cloudimage_success(capfd, delay):
     assert captured.err == ""
 
 
-def test_run_command_failure_path_not_absolute(capfd, delay):
+def test_run_command_failure_path_not_absolute(caplog, delay):
     """
     Running an invalid command on a container should croak.
     """
     runner = CliRunner()
 
-    result = runner.invoke(cli, "run -it --rm debian:buster-slim foo", catch_exceptions=False)
-    assert result.exit_code == 1
-
-    captured = capfd.readouterr()
-
-    output = strip_ansi(captured.err).strip()
-    assert output == "Failed to start transient service unit: Path foo is not absolute."
+    result = runner.invoke(cli, "run -it --rm debian:trixie-slim foo", catch_exceptions=False)
+    assert result.exit_code == 203
+    assert "Reason: foo: No such file or directory" in caplog.text
 
 
 def test_run_command_failure_command_not_found(caplog, delay):
@@ -84,10 +80,8 @@ def test_run_command_failure_command_not_found(caplog, delay):
     """
     runner = CliRunner()
 
-    result = runner.invoke(cli, "run -it --rm debian:buster-slim /bin/foo", catch_exceptions=False)
+    result = runner.invoke(cli, "run -it --rm debian:trixie-slim /bin/foo", catch_exceptions=False)
     assert result.exit_code == 203
-
-    assert "Running command in container" in caplog.text
     assert "Reason: /bin/foo: No such file or directory" in caplog.text
 
 
@@ -98,7 +92,7 @@ def test_run_stdin_stdout(monkeypatch, capsys, delay):
     """
     program_path = Path(sys.argv[0]).parent
     racker = program_path / "racker"
-    command = f"{racker} run -it --rm debian:buster-slim /bin/cat /dev/stdin"
+    command = f"{racker} run -it --rm debian:trixie-slim /bin/cat /dev/stdin"
     process = subprocess.run(shlex.split(command), input=b"foo", stdout=subprocess.PIPE, env={"TESTING": "true"})
     process.check_returncode()
     assert process.stdout == b"foo"
@@ -110,8 +104,8 @@ def test_run_stdin_stdout_original(capfd):
     runner = CliRunner()
 
     sys.stdin = io.BytesIO(b"foo")
-    result = runner.invoke(cli, "run -it --rm debian:buster-slim /bin/cat /dev/stdin", input=b"foo", catch_exceptions=False)
-    assert result.exit_code == 0
+    result = runner.invoke(cli, "run -it --rm debian:trixie-slim /bin/cat /dev/stdin", input=b"foo", catch_exceptions=False)
+    assert result.exit_code == 0, result.stderr
 
     # FIXME: Why is stdout empty?
     assert result.stdout == ""