+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 ===========
 
 - Remove support for Debian 9 "stretch"
+- Added support for Debian 13 "trixie"
 
 
 2023-02-14 0.3.0